♻️ refactor: api_schemaクレートの構造を整理し、schema.rsモジュールに分離 #3
